Display tool compatibility warning in language model selector
